backend/pyproject.toml
[tool.poetry]
name = "engicalc"
version = "0.1.0"
description = "Engineering calculation tool - Python-based calculation engine"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pint = "^0.23"
handcalcs = "^1.8.0"
python-frontmatter = "^1.0.1"
markdown-it-py = "^3.0.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
websockets = "^12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
ruff = "^0.1.6"
black = "^23.11.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
